# Fix a bug in CompareAndCopy.equals(); optimize performance