--- a/src/main/java/com/yahoo/memory/BaseWritableMemoryImpl.java
+++ b/src/main/java/com/yahoo/memory/BaseWritableMemoryImpl.java
@@ -37,16 +37,6 @@ import java.nio.ByteOrder;
  */
 abstract class BaseWritableMemoryImpl extends WritableMemory {
 
-  /**
-   * Don't use {@link sun.misc.Unsafe#copyMemory} to copy blocks of memory larger than this
-   * threshold, because internally it doesn't have safepoint polls, that may cause long
-   * "Time To Safe Point" pauses in the application. This has been fixed in JDK 9 (see
-   * https://bugs.openjdk.java.net/browse/JDK-8149596 and
-   * https://bugs.openjdk.java.net/browse/JDK-8141491), but not in JDK 8, so the Memory library
-   * should keep having this boilerplate as long as it supports Java 8.
-   */
-  static final long UNSAFE_COPY_MEMORY_THRESHOLD = 1024 * 1024;
-
   final ResourceState state;
   final Object unsafeObj; //Array objects are held here.
   final long capacity;