--- a/src/test/java/com/yahoo/memory/CopyMemoryTest.java
+++ b/src/test/java/com/yahoo/memory/CopyMemoryTest.java
@@ -5,9 +5,8 @@
 
 package com.yahoo.memory;
 
-import static com.yahoo.memory.BaseWritableMemoryImpl.UNSAFE_COPY_MEMORY_THRESHOLD;
+import static com.yahoo.memory.CompareAndCopy.UNSAFE_COPY_MEMORY_THRESHOLD;
 import static org.testng.Assert.assertEquals;
-//import static org.testng.Assert.assertTrue;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;