--- a/src/test/java/com/yahoo/memory/WritableMemoryTest.java
+++ b/src/test/java/com/yahoo/memory/WritableMemoryTest.java
@@ -10,11 +10,11 @@ import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 
-import org.testng.annotations.Test;
-
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.util.concurrent.ThreadLocalRandom;
+
+import org.testng.annotations.Test;
 
 public class WritableMemoryTest {
 
@@ -113,20 +113,19 @@ public class WritableMemoryTest {
   public void checkLargeEquals() {
     // Size bigger than UNSAFE_COPY_MEMORY_THRESHOLD; size with "reminder" = 7, to test several
     // traits of the implementation
-    byte[] bytes1 = new byte[UNSAFE_COPY_MEMORY_THRESHOLD * 2 + 7];
+    final int thresh = UNSAFE_COPY_MEMORY_THRESHOLD;
+    byte[] bytes1 = new byte[(thresh * 2) + 7];
     ThreadLocalRandom.current().nextBytes(bytes1);
     byte[] bytes2 = bytes1.clone();
     Memory mem1 = Memory.wrap(bytes1);
     Memory mem2 = Memory.wrap(bytes2);
     assertTrue(mem1.equalTo(mem2));
 
-    bytes2[UNSAFE_COPY_MEMORY_THRESHOLD + 10] =
-        (byte) (bytes1[UNSAFE_COPY_MEMORY_THRESHOLD + 10] + 1);
+    bytes2[thresh + 10] = (byte) (bytes1[thresh + 10] + 1);
     assertFalse(mem1.equalTo(mem2));
 
-    bytes2[UNSAFE_COPY_MEMORY_THRESHOLD + 10] = bytes1[UNSAFE_COPY_MEMORY_THRESHOLD + 10];
-    bytes2[UNSAFE_COPY_MEMORY_THRESHOLD * 2 + 3] =
-        (byte) (bytes1[UNSAFE_COPY_MEMORY_THRESHOLD * 2 + 3] + 1);
+    bytes2[thresh + 10] = bytes1[thresh + 10];
+    bytes2[(thresh * 2) + 3] = (byte) (bytes1[(thresh * 2) + 3] + 1);
     assertFalse(mem1.equalTo(mem2));
   }
 }