1. The fixed bug is pinpointed in the added test.

2. Performance

Before:
```
Benchmark               (size)  Mode  Cnt      Score     Error  Units
EqualsBenchmark.equals       1  avgt    5      7.992 ±   0.184  ns/op
EqualsBenchmark.equals      10  avgt    5     11.687 ±   0.643  ns/op
EqualsBenchmark.equals    1000  avgt    5    121.730 ±  11.028  ns/op
EqualsBenchmark.equals  100000  avgt    5  10674.055 ± 290.646  ns/op
```

After:
```
Benchmark               (size)  Mode  Cnt     Score     Error  Units
EqualsBenchmark.equals       1  avgt    5     7.260 ±   0.206  ns/op
EqualsBenchmark.equals      10  avgt    5    10.057 ±   0.286  ns/op
EqualsBenchmark.equals    1000  avgt    5   106.279 ±   5.983  ns/op
EqualsBenchmark.equals  100000  avgt    5  9476.547 ± 451.854  ns/op
```

It proves that safepoints and int vs long-counted loops make difference, although not as dramatic as I expected.

The benchmark is checked in https://github.com/leventov/memory-jmh-benchmarks