@leventov 2018-02-26
Whenever I fix bugs and add tests Coveralls says that the coverage decreased and "rejects" the PR. It's kind of funny.