@leerho 2018-02-26
> Coverage decreased (-0.03%) to 96.598% when pulling 62c1b2c on equals-changes into bedf853 on master.

This is a nuisance warning and I wish I knew how to set a higher tolerance than zero!  Nonetheless, coverage is a ratio. Changing either the numerator or denominator can affect the result.  So if you replace statements that were already covered with fewer covered statements the ratio of covered / total is smaller because you are subtracting the same number from both the numerator and denominator.