@leerho 2018-02-27
Thanks for finding that bug! And for the test! "Code not tested never works!"